Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@4386 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/calendar.cpp
@@ -714,7 +714,7 @@ int32_t Calendar::stringToDayNumber(const UnicodeString& string, UErrorCode& sta
 
     string.extract(0, len, number);
     number[len] = 0;
-    int32_t value = strtol(number, &end, 10); // Radix 10
+    int32_t value = (int32_t)strtol(number, &end, 10); // Radix 10
 
     if (end-number != len || len == 0 || value < 1 || value > 7)
         status = U_INVALID_FORMAT_ERROR;
